+++ libftrm/libft/ft_atof.c
@@ -6,16 +6,16 @@
 /*   By: bde-sous <bde-sous@student.42porto.com>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/02 21:59:22 by bde-sous          #+#    #+#             */
-/*   Updated: 2024/01/03 22:12:28 by bde-sous         ###   ########.fr       */
+/*   Updated: 2024/01/07 17:58:26 by bde-sous         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/libft.h"
 
-double	ft_atof(const char *str)
+float	ft_atof(const char *str)
 {
-	double	inteiro;
-	double	restante;
+	float	inteiro;
+	float	restante;
 	int		len;
 
 	inteiro = 0;
